+++ Finance-App/src/Components/Header.jsx
@@ -1,6 +1,11 @@
+import NavBar from "./NavBar.jsx"
+import "./Header_NavBar.css"
+
 const Header = () => {
     return (
-        <div class = "header"/>
+        <div class = "header">
+            <NavBar/>
+        </div>
     )
 }
 export default Header;
