+++ Finance-App/src/Pages/homepage/homepage.css
@@ -1,4 +1,4 @@
-/* App.css */
+
 .app-container {
   display: flex;
   height: 100vh;
@@ -6,28 +6,9 @@
   background-color: #f4f4f9;
 }
 
-.sidebar {
-  width: 20%;
-  background-color: #1e1e2f;
-  color: #fff;
-  padding: 20px;
-  box-shadow: 2px 0 5px rgba(0, 0, 0, 0.1);
-}
-
 .user-name {
   font-size: 1.5rem;
   margin-bottom: 20px;
-}
-
-.nav-links {
-  list-style: none;
-  padding: 0;
-}
-
-.nav-links li {
-  margin: 15px 0;
-  cursor: pointer;
-  font-size: 1.1rem;
 }
 
 .main-content {
