+++ Finance-App/src/index.css
@@ -57,12 +57,6 @@ h2 {
   text-align: center;
 }
 
-.header {
-  background-color: var(--keji-purple);
-  width: 100vw;
-  height: 25vh;
-}
-
 @media (prefers-color-scheme: light) {
   :root {
     color: #213547;
